[refactor/#580] 베스트 MR에 n회 이상 선정 시 획득 로직 최적화

See merge request s11-final/S11P31A210!245

+++ backend/src/main/java/com/reviewping/coflo/domain/mergerequest/repository/BestMrHistoryRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.repository.query.Param;
 
 public interface BestMrHistoryRepository extends CrudRepository<BestMrHistory, Long> {
     @Query("SELECT b.userId FROM BestMrHistory b GROUP BY b.userId HAVING COUNT(b) >= :count")
-    List<Long> findUsersWithAtLeastNHistory(@Param("count") int count);
+    List<Long> findUsersWithAtLeastNHistory(@Param("count") long count);
 }
